# ollama-connector/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ollama Connector API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkpoints/{id}:
    delete:
      operationId: deleteCheckpoint
      tags: [checkpoints]
      summary: Delete a checkpoint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Checkpoint deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkpoints/{id}/restore:
    post:
      operationId: restoreCheckpoint
      tags: [checkpoints]
      summary: Restore loop state from a specific checkpoint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loop restored from checkpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckpointRestoreResult"
        "404":
          description: Checkpoint not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current configuration settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update configuration settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /settings/test-connection:
    post:
      operationId: testConnection
      tags: [settings]
      summary: Test the connection to an Ollama instance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestConnectionInput"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionResult"
...
  /models/pull:
    post:
      operationId: pullModel
      tags: [models]
      summary: Pull a model from Ollama registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelPullInput"
      responses:
        "200":
          description: Pull result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPullResult"
components:
  schemas:
...
    CheckpointRestoreResult:
      type: object
      required: [checkpointId, iteration, model, prompt, state]
      properties:
        checkpointId:
          type: integer
        iteration:
          type: integer
        model:
          type: string
        prompt:
          type: string
        state:
          type: string
    Settings:
      type: object
      required: [id, ollamaUrl, defaultModel, fallbackModel, maxIterations, maxRamPercent, gitAutoCommit, enableNotifications, telegramBotToken, telegramChatId, metricsInterval, logLevel]
      properties:
        id:
          type: integer
        ollamaUrl:
          type: string
        defaultModel:
          type: string
        fallbackModel:
          type: string
        maxIterations:
          type: integer
        maxRamPercent:
          type: integer
        gitAutoCommit:
          type: boolean
        enableNotifications:
          type: boolean
        telegramBotToken:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        metricsInterval:
          type: integer
        logLevel:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
    TestConnectionInput:
      type: object
      properties:
        ollamaUrl:
          type: string
          description: URL to test (defaults to current settings)
    TestConnectionResult:
      type: object
      required: [success, reachable, ollamaUrl, message]
      properties:
        success:
          type: boolean
        reachable:
          type: boolean
        ollamaUrl:
          type: string
        message:
          type: string
        models:
          type: array
          items:
            type: object
            required: [name, size, modifiedAt, digest]
            properties:
              name:
                type: string
              size:
                type: integer
              modifiedAt:
                type: string
              digest:
                type: string
    SettingsInput:
      type: object
      properties:
        ollamaUrl:
          type: string
        defaultModel:
          type: string
        fallbackModel:
          type: string
        maxIterations:
          type: integer
        maxRamPercent:
          type: integer
        gitAutoCommit:
          type: boolean
        enableNotifications:
          type: boolean
        telegramBotToken:
          type: ["string", "null"]
        telegramChatId:
          type: ["string", "null"]
        metricsInterval:
          type: integer
        logLevel:
          type: string
          enum: [DEBUG, INFO, WARN, ERROR]
...
    ModelPullInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    ModelPullResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    DeleteResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string