# ollama-connector/lib/api-zod/src/generated/types/modelPullInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Ollama Connector API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelPullInput {
  /** @minLength 1 */
  name: string;
}