# ollama-connector/lib/api-zod/src/generated/types/testConnectionResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Ollama Connector API
 * OpenAPI spec version: 0.1.0
 */
import type { TestConnectionResultModelsItem } from './testConnectionResultModelsItem';

export interface TestConnectionResult {
  success: boolean;
  reachable: boolean;
  ollamaUrl: string;
  message: string;
  models?: TestConnectionResultModelsItem[];
}